Revert "Drop the `isBuildLess` optimization. (#1975)"
This reverts commit c5071e721ed96d0e940ea75176f2e3c1c6346458.

+++ pkg/build/build.go
@@ -512,6 +512,13 @@ func (b *Build) overlayBinSh() error {
 	return nil
 }
 
+// isBuildLess returns true if the build context does not actually do any building.
+// TODO(kaniini): Improve the heuristic for this by checking for uses/runs statements
+// in the pipeline.
+func (b *Build) isBuildLess() bool {
+	return len(b.Configuration.Pipeline) == 0
+}
+
 // getBuildConfigPURL determines the package URL for the melange config file
 // itself.
 func (b Build) getBuildConfigPURL() (*purl.PackageURL, error) {
@@ -716,67 +723,69 @@ func (b *Build) BuildPackage(ctx context.Context) error {
 	linterQueue := []linterTarget{}
 	cfg := b.workspaceConfig(ctx)
 
-	// Prepare guest directory
-	if err := os.MkdirAll(b.GuestDir, 0o755); err != nil {
-		return fmt.Errorf("mkdir -p %s: %w", b.GuestDir, err)
-	}
+	if !b.isBuildLess() {
+		// Prepare guest directory
+		if err := os.MkdirAll(b.GuestDir, 0o755); err != nil {
+			return fmt.Errorf("mkdir -p %s: %w", b.GuestDir, err)
+		}
 
-	log.Infof("building workspace in '%s' with apko", b.GuestDir)
+		log.Infof("building workspace in '%s' with apko", b.GuestDir)
 
-	guestFS := apkofs.DirFS(b.GuestDir, apkofs.WithCreateDir())
-	imgRef, err := b.buildGuest(ctx, b.Configuration.Environment, guestFS)
-	if err != nil {
-		return fmt.Errorf("unable to build guest: %w", err)
-	}
-
-	cfg.ImgRef = imgRef
-	log.Debugf("ImgRef = %s", cfg.ImgRef)
-
-	// TODO(kaniini): Make overlay-binsh work with Docker and Kubernetes.
-	// Probably needs help from apko.
-	if err := b.overlayBinSh(); err != nil {
-		return fmt.Errorf("unable to install overlay /bin/sh: %w", err)
-	}
-
-	if err := b.Runner.StartPod(ctx, cfg); err != nil {
-		return fmt.Errorf("unable to start pod: %w", err)
-	}
-	if !b.DebugRunner {
-		defer func() {
-			if err := b.Runner.TerminatePod(context.WithoutCancel(ctx), cfg); err != nil {
-				log.Warnf("unable to terminate pod: %s", err)
-			}
-		}()
-	}
-
-	// run the main pipeline
-	log.Debug("running the main pipeline")
-	pipelines := b.Configuration.Pipeline
-	if err := pr.runPipelines(ctx, pipelines); err != nil {
-		return fmt.Errorf("unable to run package %s pipeline: %w", b.Configuration.Name(), err)
-	}
-
-	for i, p := range pipelines {
-		uniqueID := strconv.Itoa(i)
-		pkg, err := p.SBOMPackageForUpstreamSource(b.Configuration.Package.LicenseExpression(), namespace, uniqueID)
+		guestFS := apkofs.DirFS(b.GuestDir, apkofs.WithCreateDir())
+		imgRef, err := b.buildGuest(ctx, b.Configuration.Environment, guestFS)
 		if err != nil {
-			return fmt.Errorf("creating SBOM package for upstream source: %w", err)
+			return fmt.Errorf("unable to build guest: %w", err)
 		}
 
-		if pkg == nil {
-			// This particular pipeline step doesn't tell us about the upstream source code.
-			continue
+		cfg.ImgRef = imgRef
+		log.Debugf("ImgRef = %s", cfg.ImgRef)
+
+		// TODO(kaniini): Make overlay-binsh work with Docker and Kubernetes.
+		// Probably needs help from apko.
+		if err := b.overlayBinSh(); err != nil {
+			return fmt.Errorf("unable to install overlay /bin/sh: %w", err)
 		}
 
-		b.SBOMGroup.AddUpstreamSourcePackage(pkg)
-	}
+		if err := b.Runner.StartPod(ctx, cfg); err != nil {
+			return fmt.Errorf("unable to start pod: %w", err)
+		}
+		if !b.DebugRunner {
+			defer func() {
+				if err := b.Runner.TerminatePod(context.WithoutCancel(ctx), cfg); err != nil {
+					log.Warnf("unable to terminate pod: %s", err)
+				}
+			}()
+		}
 
-	// add the main package to the linter queue
-	lintTarget := linterTarget{
-		pkgName:  b.Configuration.Package.Name,
-		disabled: b.Configuration.Package.Checks.Disabled,
+		// run the main pipeline
+		log.Debug("running the main pipeline")
+		pipelines := b.Configuration.Pipeline
+		if err := pr.runPipelines(ctx, pipelines); err != nil {
+			return fmt.Errorf("unable to run package %s pipeline: %w", b.Configuration.Name(), err)
+		}
+
+		for i, p := range pipelines {
+			uniqueID := strconv.Itoa(i)
+			pkg, err := p.SBOMPackageForUpstreamSource(b.Configuration.Package.LicenseExpression(), namespace, uniqueID)
+			if err != nil {
+				return fmt.Errorf("creating SBOM package for upstream source: %w", err)
+			}
+
+			if pkg == nil {
+				// This particular pipeline step doesn't tell us about the upstream source code.
+				continue
+			}
+
+			b.SBOMGroup.AddUpstreamSourcePackage(pkg)
+		}
+
+		// add the main package to the linter queue
+		lintTarget := linterTarget{
+			pkgName:  b.Configuration.Package.Name,
+			disabled: b.Configuration.Package.Checks.Disabled,
+		}
+		linterQueue = append(linterQueue, lintTarget)
 	}
-	linterQueue = append(linterQueue, lintTarget)
 
 	// run any pipelines for subpackages
 	for _, sp := range b.Configuration.Subpackages {
@@ -785,12 +794,14 @@ func (b *Build) BuildPackage(ctx context.Context) error {
 			return err
 		}
 
-		log.Infof("running pipeline for subpackage %s", sp.Name)
+		if !b.isBuildLess() {
+			log.Infof("running pipeline for subpackage %s", sp.Name)
 
-		ctx := clog.WithLogger(ctx, log.With("subpackage", sp.Name))
+			ctx := clog.WithLogger(ctx, log.With("subpackage", sp.Name))
 
-		if err := pr.runPipelines(ctx, sp.Pipeline); err != nil {
-			return fmt.Errorf("unable to run subpackage %s pipeline: %w", sp.Name, err)
+			if err := pr.runPipelines(ctx, sp.Pipeline); err != nil {
+				return fmt.Errorf("unable to run subpackage %s pipeline: %w", sp.Name, err)
+			}
 		}
 
 		// add the main package to the linter queue
@@ -932,9 +943,11 @@ func (b *Build) BuildPackage(ctx context.Context) error {
 		log.Warnf("unable to clean workspace: %s", err)
 	}
 
-	// clean build guest container
-	if err := os.RemoveAll(b.GuestDir); err != nil {
-		log.Warnf("unable to clean guest container: %s", err)
+	if !b.isBuildLess() {
+		// clean build guest container
+		if err := os.RemoveAll(b.GuestDir); err != nil {
+			log.Warnf("unable to clean guest container: %s", err)
+		}
 	}
 
 	// generate APKINDEX.tar.gz and sign it
@@ -1094,14 +1107,17 @@ func runAs(accts apko_types.ImageAccounts) string {
 
 func (b *Build) buildWorkspaceConfig(ctx context.Context) *container.Config {
 	log := clog.FromContext(ctx)
+	if b.isBuildLess() {
+		return &container.Config{
+			Arch:         b.Arch,
+			WorkspaceDir: b.WorkspaceDir,
+		}
+	}
 
-	mounts := []container.BindMount{{
-		Source:      b.WorkspaceDir,
-		Destination: container.DefaultWorkspaceDir,
-	}, {
-		Source:      "/etc/resolv.conf",
-		Destination: container.DefaultResolvConfPath,
-	}}
+	mounts := []container.BindMount{
+		{Source: b.WorkspaceDir, Destination: container.DefaultWorkspaceDir},
+		{Source: "/etc/resolv.conf", Destination: container.DefaultResolvConfPath},
+	}
 
 	if b.CacheDir != "" {
 		if fi, err := os.Stat(b.CacheDir); err == nil && fi.IsDir() {
